+++ gems/aws-sdk-s3control/lib/aws-sdk-s3control/arn/outpost_access_point_arn.rb
@@ -23,30 +23,36 @@ module Aws
       end
 
       def support_fips?
-        false
+        true
       end
 
       def validate_arn!
         unless @service == 's3-outposts'
-          raise ArgumentError, 'Must provide a valid S3 outposts ARN.'
+          raise ArgumentError, 'Must provide a valid S3 Outpost Access '\
+                               'Point ARN.'
         end
 
         if @region.empty? || @account_id.empty?
           raise ArgumentError,
-                'S3 accesspoint ARNs must contain both a region '\
-                'and an account id.'
+                'S3 Outpost Access Point ARNs must contain both a region '\
+                'and an Account ID.'
+        end
+
+        if @region.include?('-fips') || @region.include?('fips-')
+          raise ArgumentError,
+                'S3 Outpost Access Point ARNs cannot contain a FIPS region.'
         end
 
         if @type != 'outpost' && @subtype != 'accesspoint'
-          raise ArgumentError, 'Invalid ARN, resource format is not correct'
+          raise ArgumentError, 'Invalid ARN, resource format is not correct.'
         end
 
         if @outpost_id.nil? || @outpost_id.empty?
-          raise ArgumentError, 'Missing ARN outpost id.'
+          raise ArgumentError, 'Missing ARN Outpost ID.'
         end
 
         if @access_point_name.nil? || @access_point_name.empty?
-          raise ArgumentError, 'Missing ARN accesspoint name.'
+          raise ArgumentError, 'Missing ARN Access Point name.'
         end
 
         if @extra
@@ -68,11 +74,11 @@ module Aws
       end
 
       # Outpost Access Point ARNs currently do not support dualstack
-      def host_url(region, _dualstack = false, custom_endpoint = nil)
+      def host_url(region, fips = false, _dualstack = false, custom_endpoint = nil)
         if custom_endpoint
           custom_endpoint
         else
-          "s3-outposts.#{region}.amazonaws.com"
+          "s3-outposts#{'-fips' if fips}.#{region}.amazonaws.com"
         end
       end
     end